api/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
PyJWT==1.7.1
python-dotenv==0.21.0
redis==4.4.0